Create town util class
